Add files via upload
The output excel format created is correct, however it still is not extracting the info for the other columns. The overarching goal is to run this script to process the pdf downloaded from this link. The pdf is updated a few times a week and the TDOT Prequalification status As Of Date changes. The output excel file ideally will have extracted the As OF date and include that in the name of the file.